backend/requirements.txt
# Core FastAPI and server dependencies - Latest 2025 versions
fastapi==0.115.13
uvicorn[standard]==0.34.0
pydantic==2.11.7
pydantic-settings==2.9.1

# AI SDKs - Hybrid approach for production
openai==1.59.3        # For OpenAI and Groq (fully compatible)
anthropic==0.54.0     # Native for Claude (full features)
google-genai>=1.21.1   # New unified Google AI SDK (2025)

# HTTP and async dependencies  
httpx==0.28.1
anyio>=3.5.0,<5

# Configuration and environment
python-dotenv==1.0.1
PyYAML>=6.0.0

# Logging and monitoring
structlog==25.1.0

# Additional utilities
python-multipart==0.0.20
slowapi==0.1.9
bleach==6.2.0

# Search and indexing
bm25s>=0.2.0

# Supabase integration
supabase==2.15.3
PyJWT==2.9.0

# Development dependencies (uncomment for dev)
pytest==8.0.0
pytest-asyncio==0.21.0
black==24.2.0
ruff==0.1.0

